l1l2_assistant/requirements.txt
dspy-ai>=2.4.0
google-generativeai>=0.8.3
requests>=2.32.3
numpy>=1.26.4
python-dotenv>=1.0.1
